[config] move metrics_config
per title

Closes: #1722
Approved by: mimoo
