+++ config/src/config/metrics_config.rs
@@ -1,0 +1,21 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(default)]
+pub struct MetricsConfig {
+    pub dir: PathBuf,
+    pub collection_interval_ms: u64,
+}
+
+impl Default for MetricsConfig {
+    fn default() -> MetricsConfig {
+        MetricsConfig {
+            dir: PathBuf::from("metrics"),
+            collection_interval_ms: 1000,
+        }
+    }
+}
